Refactor game creation and avatar creation
